--- Content/Blueprints/Actor/Dungeon/BP_DungeonGroundItem.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f42917072b32278253a3c053dc88c60921e9c8ef983d361fecd0ad570dcc2366
-size 33264
