--- Content/Blueprints/Actor/Dungeon/BP_InteractableWeapon.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:0d297f855faccb4c3f246381e065ab80b82c6b52b3209ea5003952308b36c2b1
-size 22923
