--- Content/Blueprints/Characters/BP_RSDunAnubisCharacter.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5ae534805feef1a744f0bb2c46074bd2e42d0a66861776db68605b19f14a9ade
-size 2875
