+++ Content/Maps/MonsterControllerTestMap.umap
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:7872d143515ca45ab0dc1d407a250c0b7e2b2cb506f295b9bdbf335246b5d556
-size 139411
+oid sha256:db60aa76a1c420e45e1253b47b9a69471bc980f15b68c9c6a357e66a773534ee
+size 139020
