+++ Source/RogShop/Actor/Dungeon/GroundItem/RSDungeonGroundItem.cpp
@@ -2,7 +2,6 @@
 
 
 #include "RSDungeonGroundItem.h"
-#include "RogShop/UtilDefine.h"
 #include "Components/BoxComponent.h"
 
 // Sets default values
@@ -27,34 +26,8 @@ void ARSDungeonGroundItem::BeginPlay()
 	
 }
 
-void ARSDungeonGroundItem::InitItemInfo(FName NewDataTableKey, UStaticMesh* NewMesh)
-{
-	if (!NewMesh)
-	{
-		RS_LOG_C("Invalid StaticMesh Used", FColor::Red);
-	}
-
-	DataTableKey = NewDataTableKey;
-
-	MeshComp->SetStaticMesh(NewMesh);
-
-	MeshComp->SetCollisionEnabled(ECollisionEnabled::QueryAndPhysics);
-}
-
 void ARSDungeonGroundItem::Interact(ARSDunPlayerCharacter* Interactor)
 {
-	if (!Interactor)
-	{
-		return;
-	}
 
-	// TODO : 플레이어의 인벤토리에 아이템을 추가하는 로직 작성
-	//플레이어인벤토리컴포넌트* 변수명 = Interactor->Get플레이어인벤토리컴포넌트();
-	//if (널체크)
-	//{
-	//	플레이어 인벤토리 컴포넌트->인벤토리에 아이템 추가하는 함수 호출;
-	// 
-	//	Destroy();
-	//}
 }
