+++ Source/RogShop/Actor/Dungeon/GroundItem/RSDungeonGroundWeapon.cpp
@@ -1,43 +1,20 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 
-#include "RSInteractableWeapon.h"
+#include "RSDungeonGroundWeapon.h"
 #include "RogShop/UtilDefine.h"
 #include "Components/BoxComponent.h"
 #include "RSBaseWeapon.h"
 #include "RSDunPlayerCharacter.h"
 #include "RSPlayerWeaponComponent.h"
 
-// Sets default values
-ARSInteractableWeapon::ARSInteractableWeapon()
+ARSDungeonGroundWeapon::ARSDungeonGroundWeapon()
 {
- 	// Set this actor to call Tick() every frame.  You can turn this off to improve performance if you don't need it.
 	PrimaryActorTick.bCanEverTick = false;
 
-	SceneComp = CreateDefaultSubobject<USceneComponent>(TEXT("Scene"));
-	SetRootComponent(SceneComp);
-
-	MeshComp = CreateDefaultSubobject<UStaticMeshComponent>(TEXT("StaticMesh"));
-	MeshComp->SetupAttachment(SceneComp);
-	MeshComp->SetCollisionProfileName("Interactable");
-	MeshComp->SetCollisionEnabled(ECollisionEnabled::NoCollision);
 }
 
-// Called when the game starts or when spawned
-void ARSInteractableWeapon::BeginPlay()
-{
-	Super::BeginPlay();
-	
-}
-
-// Called every frame
-void ARSInteractableWeapon::Tick(float DeltaTime)
-{
-	Super::Tick(DeltaTime);
-
-}
-
-void ARSInteractableWeapon::InitInteractableWeapon(FName NewDataTableKey, UStaticMesh* NewMesh, const TSubclassOf<ARSDungeonItemBase> NewWeaponClass)
+void ARSDungeonGroundWeapon::InitInteractableWeapon(FName NewDataTableKey, UStaticMesh* NewMesh, const TSubclassOf<ARSDungeonItemBase> NewWeaponClass)
 {
 	if (!NewMesh)
 	{
@@ -56,7 +33,7 @@ void ARSInteractableWeapon::InitInteractableWeapon(FName NewDataTableKey, UStati
 	MeshComp->SetCollisionEnabled(ECollisionEnabled::QueryAndPhysics);
 }
 
-void ARSInteractableWeapon::Interact(ARSDunPlayerCharacter* Interactor)
+void ARSDungeonGroundWeapon::Interact(ARSDunPlayerCharacter* Interactor)
 {
 	if (!Interactor)
 	{
