+++ Source/RogShop/RogShop.Build.cs
@@ -24,8 +24,9 @@ public class RogShop : ModuleRules
             Path.Combine(ModuleDirectory, "CheatManager"),
             Path.Combine(ModuleDirectory, "GameModeBase"),
             Path.Combine(ModuleDirectory, "GameInstance"),
-            Path.Combine(ModuleDirectory, "Actor", "Dungeon", "Weapon"),
             Path.Combine(ModuleDirectory, "Actor", "Dungeon"),
+            Path.Combine(ModuleDirectory, "Actor", "Dungeon", "Weapon"),
+            Path.Combine(ModuleDirectory, "Actor", "Dungeon", "GroundItem"),
             Path.Combine(ModuleDirectory, "Interface"),
             Path.Combine(ModuleDirectory, "DataTable"),
             Path.Combine(ModuleDirectory, "Object", "Relic"),
